+++ Gruntfile.js
@@ -181,7 +181,7 @@ module.exports = function( grunt ) {
         var clientConfigPath = "public/temp-client-config.json";
         if ( task === 'create' ) {
             var config = require( './app/models/config-model' );
-            grunt.file.write( clientConfigPath, JSON.stringify( config.client() ) );
+            grunt.file.write( clientConfigPath, JSON.stringify( config.client ) );
             grunt.log.writeln( 'File ' + clientConfigPath + ' created' );
         } else if ( task === 'remove' ) {
             grunt.file.delete( clientConfigPath );