+++ test/server/survey-model.spec.js
@@ -10,7 +10,7 @@ var model,
     expect = chai.expect,
     chaiAsPromised = require( "chai-as-promised" ),
     redis = require( "redis" ),
-    config = require( "../../config/config" ),
+    config = require( "../../app/models/config-model" ).server,
     model = require( '../../app/models/survey-model' ),
     client = redis.createClient( config.redis.main.port, config.redis.main.host, {
         auth_pass: config.redis.main.password